update nodes check
