+++ intent_client/src/config/api.ts
@@ -1,8 +1,8 @@
 export const API_CONFIG = {
     BASE_URL: "http://127.0.0.1:8001/api/v1",
     ENDPOINTS: {
-        MESSAGE_VERIFY: "/messages/verify",
-        MESSAGE_CONFIRM: "/messages/conversation/confirm",
+        MESSAGE_VERIFY: "/intents/verify",
+        MESSAGE_CONFIRM: "/intents/conversation/confirm",
         TOPO_BUILD: "/topo/build",
         TOKEN: "/generate-token",
     },
