+++ intent_client/src/features/intentChat/services/api.ts
@@ -2,7 +2,7 @@ import {IntentMessageDTO} from "../types";
 import {API_CONFIG} from "../../../config/api.ts";
 import axios from "axios";
 
-export const sendMessage = async (token: string, intentMessageDTO: IntentMessageDTO): Promise<IntentMessageDTO> => {
+export const sendMessage = async (token: string, intentMessageDTO: IntentMessageDTO): Promise<any> => {
     const response = await axios.post<IntentMessageDTO>(
         `${API_CONFIG.BASE_URL}${API_CONFIG.ENDPOINTS.MESSAGE_VERIFY}`,
         intentMessageDTO,
